requirements.txt
Cheetah==2.4.4
Django==1.9.8
Jinja2==2.9.6
Landscape-Client==14.12
MarkupSafe==1.0
PAM==0.4.2
PyYAML==3.10
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
Unidecode==0.04.21
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
dj-database-url==0.4.2
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-bootstrap3==9.0.0
django-extensions==1.8.1
django-nvd3==0.9.7
django-user-accounts==1.3.1
djangorestframework==3.6.3
gyp==0.1
html5lib==0.999
jsonfield==2.0.2
jsonpatch==1.3
jsonpointer==1.0
numpy==1.12.1
oauth==1.0.1
pinax-eventlog==1.1.2
pinax-theme-bootstrap==7.10.1
pinax-webanalytics==2.0.4
prettytable==0.7.2
pyOpenSSL==0.13
pycurl==7.19.3
pyserial==2.6
python-apt==0.9.3.5ubuntu2
python-dateutil==2.6.0
python-debian==0.1.21-nmu2ubuntu2
python-memcached==1.58
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2016.10
requests==2.2.1
six==1.5.2
ssh-import-id==3.21
urllib3==1.7.1
wheel==0.24.0
whitenoise==3.3.0
wsgiref==0.1.2
zope.interface==4.0.5
